+++ packages/expo-modules-core/android/src/main/java/expo/modules/interfaces/barcodescanner/BarCodeScannerInterface.java
@@ -1,10 +1,10 @@
-package org.unimodules.interfaces.barcodescanner;
+package expo.modules.interfaces.barcodescanner;
 
 import android.graphics.Bitmap;
 
 import java.util.List;
 
-public interface BarCodeScanner {
+public interface BarCodeScannerInterface {
   BarCodeScannerResult scan(byte[] imageData, int width, int height, int rotation);
   List<BarCodeScannerResult> scanMultiple(Bitmap bitmap);
   void setSettings(BarCodeScannerSettings settings);